+++ common/src/main/java/com/chefmooon/frightsdelight/common/block/FrightsDelightBushBlock.java
@@ -65,10 +65,14 @@ public class FrightsDelightBushBlock extends BushBlock implements BonemealableBl
                 double d = Math.abs(entity.getX() - entity.xOld);
                 double e = Math.abs(entity.getZ() - entity.zOld);
                 if (d >= HURT_SPEED_THRESHOLD || e >= HURT_SPEED_THRESHOLD) {
-                    entity.hurt(DamageSource.SWEET_BERRY_BUSH, 1.0F); // TODO: custom damage source here?
+                    entity.hurt(getDamageSource(), 1.0F); // TODO: custom damage source here?
                 }
             }
         }
+    }
+
+    public DamageSource getDamageSource() {
+        return DamageSource.SWEET_BERRY_BUSH;
     }
 
     @Override