+++ common/src/main/java/com/chefmooon/frightsdelight/common/block/SoulBerryBushBlock.java
@@ -1,6 +1,6 @@
 package com.chefmooon.frightsdelight.common.block;
 
-import com.chefmooon.frightsdelight.common.registry.FrightsDelightBlocks;
+import com.chefmooon.frightsdelight.common.registry.FrightsDelightDamageTypes;
 import com.chefmooon.frightsdelight.common.registry.FrightsDelightItems;
 import com.chefmooon.frightsdelight.common.tag.FrightsDelightTags;
 import com.chefmooon.frightsdelight.common.utility.HolderLookupHelper;
@@ -13,6 +13,7 @@ import net.minecraft.tags.TagKey;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
+import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
@@ -104,6 +105,11 @@ public class SoulBerryBushBlock extends FrightsDelightBushBlock {
     public void performBonemeal(ServerLevel level, RandomSource random, BlockPos pos, BlockState state) {
         updateConditions(state, level, pos);
         super.performBonemeal(level, random, pos, state);
+    }
+
+    @Override
+    public DamageSource getDamageSource() {
+        return FrightsDelightDamageTypes.SOUL_BERRY_BUSH;
     }
 
     @Override