+++ changelog.md
@@ -8,6 +8,10 @@ This mod is compatible with Minecraft 1.19.2 and Farmer's Delight Forge and Fabr
 
 Beta release for Minecraft 1.19.2!
 
+### Added
+
+- Custom damage source and death messages for Soul and Wither Berry Bushes
+
 ### Known Bugs
 - Punch Bowls break speed is not faster with knife (Fabric only)